src/main/java/be/ucll/david/Tasks/dto/UserDTO.java
package be.ucll.david.Tasks.dto;

import be.ucll.david.Tasks.model.UserRole;

import javax.validation.constraints.NotEmpty;
import javax.validation.constraints.NotNull;

public class UserDTO {
    @NotNull
    private Long id;
    @NotEmpty
    private String username;
    @NotNull
    private UserRole role;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public UserRole getRole() {
        return role;
    }

    public void setRole(UserRole role) {
        this.role = role;
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }
}
